Fixed an error getting user location on startup.
The game would crash if a user logged in who was located in the void
when they last logged out (because the value 'None' was getting saved to
the database). The userize function now does better checking
to make sure that it only tries to set a the user's location if the user
actually has a valid one saved.