Added TextEditMode, fixed some bugs.
Added TextEditMode for easier editing of multi-line descriptions for
areas, npcs, room, and items. Also gave every object a save function so
that it could save itself, rather than having to call the update_dict
and insert_dict functions for the db. Also fixed the Destroy command so
that it properly destroys things, even areas.