Added the Script object model, fixed TextEditMode.
We now have a Script object model! Scripts can be created as part of an
area during BuildMode, as well as be edited, saved, loaded, etc. Also
fixed a bug in TextEditMode that would not recognize a line of that line
was considered to be unfinished (didn't end in a ., !, or ? symbol).