Made the MUD telnet friendly!
Before the player gets passed from the connection handler to world, we
attempt to initiate line_mode. This fixes the problem with clients like
putty that expect to send each character as it's entered. We also make
sure that every newline is accompanied by a line feed, and that we
ignore any IAC commands (as per the RFC for implementing the minimal
telnet protocol).