Changed the naming scheme for BuildMode accessors.
The attributes for BuildMode objects are usually edited using the
keywords add, remove, and set. The naming scheme (before the change) was
that every function defined in the models that was prefaced with set_,
add_, or remove_ was accessible via the BuildMode commands Set, Add, and
Remove. Now these functions must be prefaced with build_add_,
build_remove, and build_set to be accessible via the Set, Add, and
Remove BuildCommands.