Fixed portal bug
Broken portals were caused by an import bug which relied on
calling older code*. Changes did not affect the shiny import
format in any way, just the methods by which item_types
are loaded into the game.

*old code was based on the user method 'build_add_type'
which once allowed you to pass it a pre-built item type.
Later, we decided to standerdize the 'build add' command
for users, which also broke imports. The fix was to add
a separate method for importing item types.